requirements.txt
requests~=2.25.1
vyper-config~=1.0